requirements.txt
pandas
geopandas
folium
streamlit_folium
plotly.express
seaborn
